Updated the program with the code to allow Simulog to run Ashita client sessions remotely without having this program be in the Ashita program.
In addition, new code was added so that the starting directory for the Open File Dialog box will be saved to the last directory the user selected a file in. This makes selecting the next configuration file or client easier to navigate to.

The code has also been adjusted to only create new configuration files in same folder as the selected configuration file. They will no longer be copied locally into Simulog's folder anymore.